requirements.txt
python-dotenv
requests
openai>=1.30.0
# For Bitunix (unofficial but works great)
ccxt>=4.0.0